requirements.txt
pandas==1.5.3
pickle5==0.0.11
tensorboard
datasets
transformers[torch]
sentencepiece
tiktoken
scikit-learn
llama_cpp_python==0.2.28
#transformers==4.19.2
#torch==1.11.0